убери мусор бичара!
